+++ src/ai/backend/manager/services/agent/actions/base.py
