+++ src/ai/backend/manager/services/agent/actions/get_watcher_status.py
@@ -3,7 +3,7 @@ from typing import Any, Optional, override
 
 from ai.backend.common.types import AgentId
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.agent.base import AgentAction
+from ai.backend.manager.services.agent.actions.base import AgentAction
 
 
 @dataclass
