+++ src/ai/backend/manager/services/keypair_resource_policy/actions/create_keypair_resource_policy.py
@@ -4,7 +4,9 @@ from typing import Any, Optional, override
 from ai.backend.manager.actions.action import BaseActionResult
 from ai.backend.manager.data.resource.types import KeyPairResourcePolicyData
 from ai.backend.manager.models.resource_policy import KeyPairResourcePolicyRow
-from ai.backend.manager.services.keypair_resource_policy.base import KeypairResourcePolicyAction
+from ai.backend.manager.services.keypair_resource_policy.actions.base import (
+    KeypairResourcePolicyAction,
+)
 from ai.backend.manager.types import OptionalState, TriState
