+++ src/ai/backend/manager/services/project_resource_policy/actions/create_project_resource_policy.py
@@ -6,7 +6,9 @@ from ai.backend.manager.data.resource.types import ProjectResourcePolicyData
 from ai.backend.manager.models.resource_policy import (
     ProjectResourcePolicyRow,
 )
-from ai.backend.manager.services.project_resource_policy.base import ProjectResourcePolicyAction
+from ai.backend.manager.services.project_resource_policy.actions.base import (
+    ProjectResourcePolicyAction,
+)
 from ai.backend.manager.types import OptionalState
