+++ src/ai/backend/manager/services/resource_preset/actions/base.py
