+++ src/ai/backend/manager/services/resource_preset/actions/check_presets.py
@@ -4,7 +4,7 @@ from typing import Any, Mapping, Optional, override
 
 from ai.backend.common.types import ResourceSlot
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.resource_preset.base import ResourcePresetAction
+from ai.backend.manager.services.resource_preset.actions.base import ResourcePresetAction
 
 
 @dataclass
