+++ src/ai/backend/manager/services/resource_preset/actions/list_presets.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import Any, Optional, override
 
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.services.resource_preset.base import ResourcePresetAction
+from ai.backend.manager.services.resource_preset.actions.base import ResourcePresetAction
 
 
 @dataclass
