+++ src/ai/backend/manager/services/user_resource_policy/actions/create_user_resource_policy.py
@@ -6,7 +6,7 @@ from ai.backend.manager.data.resource.types import UserResourcePolicyData
 from ai.backend.manager.models.resource_policy import (
     UserResourcePolicyRow,
 )
-from ai.backend.manager.services.user_resource_policy.base import UserResourcePolicyAction
+from ai.backend.manager.services.user_resource_policy.actions.base import UserResourcePolicyAction
 from ai.backend.manager.types import OptionalState
