+++ src/ai/backend/manager/services/user_resource_policy/actions/modify_user_resource_policy.py
@@ -3,7 +3,7 @@ from typing import Any, Optional, override
 
 from ai.backend.manager.actions.action import BaseActionResult
 from ai.backend.manager.data.resource.types import UserResourcePolicyData
-from ai.backend.manager.services.user_resource_policy.base import UserResourcePolicyAction
+from ai.backend.manager.services.user_resource_policy.actions.base import UserResourcePolicyAction
 from ai.backend.manager.types import OptionalState
